# backend/cmd/api/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing courses, modules, profiles, and user progress in a
    learning management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Course Management System API
  version: "1.0"
paths:
  /admin/factory-reset:
    post:
      consumes:
      - application/json
      description: Clear all data from the database (use with caution!)
      produces:
      - application/json
      responses:
        "200":
          description: Database reset successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Factory reset database
      tags:
      - admin
  /admin/stats:
    get:
      consumes:
      - application/json
      description: Retrieve basic statistics about the database (counts of profiles,
        courses, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: Database statistics
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get database statistics
      tags:
      - admin
...
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"